+++ src/command_modules/azure-cli-sqlvm/azure/cli/command_modules/sqlvm/_help.py
@@ -69,17 +69,17 @@ helps['sql vm create'] = """
             az sql vm create -n sqlvm -g myresourcegroup -l eastus --image-sku Enterprise --license-type AHUB
         - name: Enable R services in SQL2016 onwards.
           text: >
-            az sql vm create -n sqlvm -g myresourcegroup -l eastus --enable-r-services true
+            az sql vm create -n sqlvm -g myresourcegroup -l eastus --license-type PAYG --enable-r-services true
         - name: Create SQL virtual machine and configure auto backup settings.
           text: >
-            az sql vm create -n sqlvm -g myresourcegroup -l eastus --backup-schedule-type manual --full-backup-frequency Weekly --full-backup-start-hour 2 --full-backup-duration 2
+            az sql vm create -n sqlvm -g myresourcegroup -l eastus --license-type PAYG --backup-schedule-type manual --full-backup-frequency Weekly --full-backup-start-hour 2 --full-backup-duration 2
             --sa-key {storageKey} --storage-account 'https://storageacc.blob.core.windows.net/' --retention-period 30 --log-backup-frequency 60
         - name: Create SQL virtual machine and configure auto patching settings.
           text: >
-            az sql vm create -n sqlvm -g myresourcegroup -l eastus --day-of-week sunday --maintenance-window-duration 60 --maintenance-window-start-hour 2
+            az sql vm create -n sqlvm -g myresourcegroup -l eastus --license-type PAYG --day-of-week sunday --maintenance-window-duration 60 --maintenance-window-start-hour 2
         - name: Create SQL virtual machine and configure SQL connectivity settings.
           text: >
-            az sql vm create -n sqlvm -g myresourcegroup -l eastus --connectivity-type private --port 1433 --sql-auth-update-username {newlogin} --sql-auth-update-pwd {sqlpassword}
+            az sql vm create -n sqlvm -g myresourcegroup -l eastus --license-type PAYG --connectivity-type private --port 1433 --sql-auth-update-username {newlogin} --sql-auth-update-pwd {sqlpassword}
     """
 helps['sql vm update'] = """
     type: command