+++ src/command_modules/azure-cli-sqlvm/azure/cli/command_modules/sqlvm/_params.py
@@ -162,7 +162,7 @@ def load_arguments(self, _):
     with self.argument_context('sql vm', arg_group='SQL Server License') as c:
         c.argument('sql_server_license_type',
                    help='SQL Server license type.',
-                   options_list=['--license-type'],
+                   options_list=['--license-type', '-b'],
                    arg_type=get_enum_type(SqlServerLicenseType))
         c.argument('sql_image_sku',
                    options_list=['--image-sku'],